requirements.txt
Pillow
future
matplotlib
numpy
tensorboard>=1.14
torch>=1.2
torchvision>=0.4
PySuperTuxKart
opencv-python
ffmpeg-python
